\name{O2DIAdyna}
\alias{O2DIAdyna}
\title{
  Dynamic solution of O2 dynamics in the sediment.
}
\description{
  \code{O2DIAdyna} dynamically runs the O2DIA model.
}
\usage{
  O2DIAdyna (parms = list(), times = seq(0, 1, length.out = 10), 
     yini = NULL, Grid = NULL, 
     Temperature = 10, PAR0 = 1000, Waterheight = 0.01,
     Tfunction = "Q10", verbose = FALSE)
  }
\arguments{

  \item{parms}{ A list with parameter values, Available parameters can be listed using function \link{O2DIAparms}.

    See details.
  }

  \item{yini }{Initial guess of the steady-state solution.
  }
  
 \item{Grid}{ If specified: either an object, as returned by \code{setup.grid.1D} from the package \code{ReacTran}, 
   a vector of length 301 with the transport distances (from mid to mid of layers, upper interface = diffusive boundary layer), 
   or one number with the constant layer thickness. 
   If \code{NULL}, it is defined as \code{setup.grid.1D(x.up = -x_DBL, dx.1 = c( x_DBL/20, 1e-6), N = c(20, N-20), L = c(DBL, x_down))}, where x_DBL and x_down are parameters defining the thickness of the diffusive boundary layer and the length of the modelede sediment. This means there are 20 layers defined in the diffusive boundary layer, the remainder in the sediment, with initial thickness = 1e-6 m, and increasing with depth for 300-20 layers.  
  } 
  \item{times}{ the time for which to estimate the steady-state condition; can be one value, or a vector (in which case only the first value will be used).
  }
  \item{Temperature}{ the temperature forcing function, units of [dgC]. Can be one value, a function that takes time as input, or a two-columned data.frame or matrix. In the latter case, it is assumed that the first column represents time, and the data series will be used to approximate the value for the input \code{times}. When a function, then the value will be generated by calling the function for the input \code{times}.
  }
  \item{PAR0}{  forcing function for the photosynthetically active radiation at the sediment-water  interface, units of [uEinst/m2/s]. Can be one value, a function that takes time as input, or a two-columned data.frame or matrix. In the latter case, it is assumed that the first column represents time, and the data series will be used to approximate the value for the input \code{times}. When a function, then the value will be generated by calling the function for the input \code{times}.
  }  
  \item{Waterheight}{  forcing function for water height, units of [m]. This has an effect on the sediment-water exchange fluxes based on water height being > 0 or = 0. Can be one value, a function that takes time as input, or a two-columned data.frame or matrix. In the latter case, it is assumed that the first column represents time, and the data series will be used to approximate the value for the input \code{times}. When a function, then the value will be generated by calling the function for the input \code{times}.
  }
  \item{Tfunction}{ How the temperature effect on rates needs to be estimated. When "Q10", a Q10-formulation (exponential) will be used, as driven by parameters \code{Q10_min, Q10_odureox, Q10_mpbprod, Q10_basalresp, Q10_photoresp}. If a different name, a cardinal function will be used, driven by parameters \code{Tmax_MPBprod, Tmin_MPBprod, Topt_MPBprod, Tmax_min, Tmin_min, Topt_min, Tmax_reox, Tmin_reox, Topt_reox, Tmax_basalresp, Tmin_basalresp, Topt_basalresp, Tmax_photoresp, Tmin_photoresp, Topt_photoresp}
  }
  \item{verbose}{ If \code{TRUE}, will write to the screen.}
}

\value{
 A matrix of class \code{O2DIAdyn} and \code{deSolve}, as generated by the solver from R-package \code{deSolve} (\link{ode.1D}).
 
It contains several output columns, the first is \code{time}. 

The other columns contain, for all time values the following:
 \itemize{
   \item \code{O2, ODU}, the state variables in all 300 grid points, i.e. comprising 600 columns.
   \item \code{O2SWIflux, O2deepflux, ODUSWIflux, ODUdeepflux}, the sediment-water interface (SWI) and deep (burial) fluxes, in mol O2/m2/d.
   \item \code{O2consumption, O2prod, O2min, O2odureox, O2basalresp, O2photoresp, ODUconsumption}, the vertical profiles of total oxygen consumption, of oxygen production, of growth respiration+mineralistion, of ODU reoxidation, basal respiration, photorespiration and ODU reoxidation, all in mol O2/m3 liquid/d.
   \item \code{rODUreox, rMPBprod, rbasalresp} the actual rates of processes (temperature dependent).
   \item \code{Biodist}, the values of microphytobenthos biomass, in molC/m3solid (300 elements)
   \item \code{Light, kext, PI}, the light profile (uEinst/m2/s), the light extinction coefficient profile (/m), the relative photosynthesis (light limitation), in all 300 layers (each 300 columns).
    \item \code{Totalcons, Totalmin, Totalbasalresp, Totalphotoresp, Totalodureox, Totalprod}, the vertically integrated rates of total oxygen consumption, of growth respiration+mineralistion, of basal respiration, of photorespiration, of ODU reoxidation, and of oxygen production, all in mol O2/m2/d.
    \item \code{Temperature, PAR0, Waterheight}, the forcing functions.
  }

 
 The meaning and units of these columns can be assessed via the R-functions:
 
 \code{O2DIAsvar()}, \code{O2DIA1D()}, \code{O2DIA0D()}. See \link{O2DIA0D}.

}

\author{
  Karline Soetaert
}

\note{
  The model application starts by estimating the steady-state condition
  of the model, in case \code{yini} is not provided. This steady-state condition is then used as a starting
  condition for a dynamic simulation.

 As the model is implemented in fortran, a warning may be issued at the start, about the used integration method - this warning can be ignored.

}


\examples{


#===========================================
# Show parameter values
#===========================================

  O2DIAparms()

#===========================================
# Default run
#===========================================

  times <- seq(0, to = 1, by=60/86400)  # output every minute
  out   <- O2DIAdyna(times       = times, 
                     Temperature = 20, 
                     PAR0        = 636)
  
  image2D(out)
    
  O2DIAbudgetO2(out)  

#===========================================
# forcing functions data sets
#===========================================
  
  Light  <- pmax(0, 2000 * sin(times*2*pi - 0.5*pi))
  Temp   <- 20 + 2*sin(times*2*pi - pi)
  
  out2 <- O2DIAdyna(times       = times, 
                    Temperature = cbind(times, Temp), 
                    PAR0        = cbind(times, Light))

#===========================================
# forcing functions
#===========================================

  Light_F  <- function (t) pmax(0, 1000 * sin(t*2*pi - 0.5*pi))
  Temp_F   <- function (t) 25 + 2*sin(t*2*pi - pi)
  
  out3 <- O2DIAdyna(times       = times, 
                    Temperature = Temp_F, 
                    PAR0        = Light_F)

  plot(out3, out2, 
       which = c("O2SWIflux", "Temperature", "PAR0"))                     

#===========================================
# Altered parameter values
#===========================================

  parms <- list(Q10_mpbprod = 2, 
                Q10_basalresp = 2)  
  out4 <- O2DIAdyna(parms  = parms, 
                    times       = times, 
                    Temperature = Temp_F, 
                    PAR0        = Light_F)
  
  plot(out3, out4, 
       which = c("O2SWIflux", "Temperature", "PAR0"))                     

}
\references{

to be filled
}

\keyword{ utilities }
